packages/call-oci-vision-fn/requirements.txt
oci
python-dotenv
opencv-python